# backend/app/api/routes/series.py
from __future__ import annotations

from pathlib import Path

from fastapi import APIRouter, Depends, HTTPException
from fastapi.responses import FileResponse

from app.api.deps import get_storage, get_study_service
from app.api.schemas import InstanceItem, InstanceListResponse, SeriesSummary
from app.infra.storage import StorageService
from app.services.study_service import StudyService

router = APIRouter(prefix="/series", tags=["series"])


@router.get("/{series_uid}", response_model=SeriesSummary)
def get_series(series_uid: str, svc: StudyService = Depends(get_study_service)) -> SeriesSummary:
    series = svc.get_series(series_uid)
    if series is None:
        raise HTTPException(status_code=404, detail={"code": "SERIES_NOT_FOUND", "message": series_uid})
    return SeriesSummary.model_validate(svc.series_to_dict(series))


@router.get("/{series_uid}/instances", response_model=InstanceListResponse)
def list_instances(series_uid: str, svc: StudyService = Depends(get_study_service)) -> InstanceListResponse:
    series = svc.get_series(series_uid)
    if series is None:
        raise HTTPException(status_code=404, detail={"code": "SERIES_NOT_FOUND", "message": series_uid})
    instances = svc.list_instances(series_uid)
    items = [
        InstanceItem(
            sop_uid=i.sop_uid,
            instance_number=i.instance_number,
            rows=i.rows,
            cols=i.cols,
            frame_url=f"/api/v1/series/{series_uid}/frames/{idx}",
            wadouri=f"wadouri:/api/v1/series/{series_uid}/frames/{idx}",
        )
        for idx, i in enumerate(instances)
    ]
    return InstanceListResponse(
        items=items,
        total=len(items),
        series=SeriesSummary.model_validate(svc.series_to_dict(series)),
    )


@router.get("/{series_uid}/frames/{idx}")
def get_frame(
    series_uid: str,
    idx: int,
    svc: StudyService = Depends(get_study_service),
) -> FileResponse:
    instances = svc.list_instances(series_uid)
    if idx < 0 or idx >= len(instances):
        raise HTTPException(status_code=404, detail={"code": "FRAME_NOT_FOUND", "message": f"index {idx}"})
    path = Path(instances[idx].file_path)
    if not path.exists():
        raise HTTPException(status_code=404, detail={"code": "FRAME_FILE_MISSING", "message": str(path)})
    return FileResponse(
        path,
        media_type="application/dicom",
        filename=path.name,
        headers={"Accept-Ranges": "bytes"},
    )


@router.get("/{series_uid}/thumbnail")
def get_thumbnail(
    series_uid: str,
    svc: StudyService = Depends(get_study_service),
    _storage: StorageService = Depends(get_storage),
) -> FileResponse:
    series = svc.get_series(series_uid)
    if series is None:
        raise HTTPException(status_code=404, detail={"code": "SERIES_NOT_FOUND", "message": series_uid})
    if not series.thumbnail_path:
        raise HTTPException(status_code=404, detail={"code": "THUMB_MISSING", "message": series_uid})
    path = Path(series.thumbnail_path)
    if not path.exists():
        raise HTTPException(status_code=404, detail={"code": "THUMB_MISSING", "message": series_uid})
    return FileResponse(path, media_type="image/png")