# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoxFlow",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/studies/upload": {
      "post": {
        "tags": [
          "studies"
        ],
        "summary": "Upload Studies",
        "operationId": "upload_studies_api_v1_studies_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_studies_api_v1_studies_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/studies": {
      "get": {
        "tags": [
          "studies"
        ],
        "summary": "List Studies",
        "operationId": "list_studies_api_v1_studies_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "modality",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Modality"
            }
          },
          {
            "name": "body_part",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Body Part"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_StudySummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/studies/{study_uid}": {
      "get": {
        "tags": [
          "studies"
        ],
        "summary": "Get Study",
        "operationId": "get_study_api_v1_studies__study_uid__get",
        "parameters": [
          {
            "name": "study_uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Study Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudySummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/series/{series_uid}": {
      "get": {
        "tags": [
          "series"
        ],
        "summary": "Get Series",
        "operationId": "get_series_api_v1_series__series_uid__get",
        "parameters": [
          {
            "name": "series_uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Series Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/series/{series_uid}/instances": {
      "get": {
        "tags": [
          "series"
        ],
        "summary": "List Instances",
        "operationId": "list_instances_api_v1_series__series_uid__instances_get",
        "parameters": [
          {
            "name": "series_uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Series Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/series/{series_uid}/frames/{idx}": {
      "get": {
        "tags": [
          "series"
        ],
        "summary": "Get Frame",
        "operationId": "get_frame_api_v1_series__series_uid__frames__idx__get",
        "parameters": [
          {
            "name": "series_uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Series Uid"
            }
          },
          {
            "name": "idx",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idx"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/series/{series_uid}/thumbnail": {
      "get": {
        "tags": [
          "series"
        ],
        "summary": "Get Thumbnail",
        "operationId": "get_thumbnail_api_v1_series__series_uid__thumbnail_get",
        "parameters": [
          {
            "name": "series_uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Series Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Models",
        "operationId": "list_models_api_v1_models_get",
        "parameters": [
          {
            "name": "enabled_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Enabled Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/models/{model_id}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Model",
        "operationId": "get_model_api_v1_models__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSpec"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "models"
        ],
        "summary": "Patch Model",
        "operationId": "patch_model_api_v1_models__model_id__patch",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSpec"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_studies_api_v1_studies_upload_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_studies_api_v1_studies_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InstanceItem": {
        "properties": {
          "sop_uid": {
            "type": "string",
            "title": "Sop Uid"
          },
          "instance_number": {
            "type": "integer",
            "title": "Instance Number"
          },
          "rows": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows"
          },
          "cols": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cols"
          },
          "frame_url": {
            "type": "string",
            "title": "Frame Url"
          },
          "wadouri": {
            "type": "string",
            "title": "Wadouri"
          }
        },
        "type": "object",
        "required": [
          "sop_uid",
          "instance_number",
          "frame_url",
          "wadouri"
        ],
        "title": "InstanceItem"
      },
      "InstanceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InstanceItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "series": {
            "$ref": "#/components/schemas/SeriesSummary"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "series"
        ],
        "title": "InstanceListResponse"
      },
...
      "ModelListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ModelSpec"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ModelListResponse"
      },
...
      "ModelPatchRequest": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "default_params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Params"
          }
        },
        "type": "object",
        "title": "ModelPatchRequest"
      },
...
      "ModelSpec": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "task_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "segmentation",
                  "detection",
                  "classification"
                ]
              },
              {
                "type": "string"
              }
            ],
            "title": "Task Type"
          },
          "modalities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modalities"
          },
          "body_parts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Body Parts"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "input_constraints": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Constraints"
          },
          "params_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params Schema"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/ModelOutputSpec"
            },
            "type": "array",
            "title": "Outputs"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Metrics"
          },
          "expected_latency_ms": {
            "type": "integer",
            "title": "Expected Latency Ms",
            "default": 0
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "default_params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Default Params"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "task_type",
          "modalities",
          "body_parts",
          "description"
        ],
        "title": "ModelSpec"
      },
...
      "Page_StudySummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudySummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[StudySummary]"
      },
...
      "SeriesSummary": {
        "properties": {
          "series_uid": {
            "type": "string",
            "title": "Series Uid"
          },
          "study_uid": {
            "type": "string",
            "title": "Study Uid"
          },
          "modality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modality"
          },
          "body_part": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Part"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "series_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Number"
          },
          "rows": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows"
          },
          "cols": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cols"
          },
          "num_instances": {
            "type": "integer",
            "title": "Num Instances",
            "default": 0
          },
          "spacing": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spacing"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "series_uid",
          "study_uid"
        ],
        "title": "SeriesSummary"
      },
      "StudySummary": {
        "properties": {
          "study_uid": {
            "type": "string",
            "title": "Study Uid"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "patient_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Name"
          },
          "patient_sex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Sex"
          },
          "patient_age": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Age"
          },
          "study_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Date"
          },
          "study_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Description"
          },
          "modality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modality"
          },
          "body_part": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Part"
          },
          "institution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institution"
          },
          "num_series": {
            "type": "integer",
            "title": "Num Series",
            "default": 0
          },
          "num_instances": {
            "type": "integer",
            "title": "Num Instances",
            "default": 0
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "series": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SeriesSummary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series"
          }
        },
        "type": "object",
        "required": [
          "study_uid"
        ],
        "title": "StudySummary"
      },
      "StudyUploadResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudySummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "StudyUploadResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}